+++ pyfakefs/tests/fake_filesystem_unittest_test.py
@@ -167,10 +167,22 @@ class TestPatchingImports(TestPyfakefsUnittestBase):
         self.assertTrue(
             pyfakefs.tests.import_as_example.check_if_exists5(file_path))
 
+    def test_import_function_from_os_path_as_other_name(self):
+        file_path = '/foo/bar'
+        self.fs.create_dir(file_path)
+        self.assertTrue(
+            pyfakefs.tests.import_as_example.check_if_exists6(file_path))
+
     def test_import_function_from_os(self):
         file_path = '/foo/bar'
         self.fs.create_file(file_path, contents=b'abc')
-        stat_result = pyfakefs.tests.import_as_example.file_stat(file_path)
+        stat_result = pyfakefs.tests.import_as_example.file_stat1(file_path)
+        self.assertEqual(3, stat_result.st_size)
+
+    def test_import_function_from_os_as_other_name(self):
+        file_path = '/foo/bar'
+        self.fs.create_file(file_path, contents=b'abc')
+        stat_result = pyfakefs.tests.import_as_example.file_stat2(file_path)
         self.assertEqual(3, stat_result.st_size)