+++ pyfakefs/tests/import_as_example.py
@@ -14,10 +14,12 @@
 Example module that is used for testing modules that import file system modules
 to be patched under another name.
 """
-from os import path
-from os.path import exists
 import os as my_os
+from os import path
 from os import stat
+from os import stat as my_stat
+from os.path import exists
+from os.path import exists as my_exists
 
 import sys
 
@@ -56,9 +58,19 @@ def check_if_exists5(filepath):
     return exists(filepath)
 
 
-def file_stat(filepath):
+def check_if_exists6(filepath):
+    # tests patching `exists` imported from os.path as other name
+    return my_exists(filepath)
+
+
+def file_stat1(filepath):
     # tests patching `stat` imported from os
     return stat(filepath)
+
+
+def file_stat2(filepath):
+    # tests patching `stat` imported from os as other name
+    return my_stat(filepath)
 
 
 def system_stat(filepath):